v3: relocate WpdsThemeProvider into core-default engine
Move src/runtime/styles/WpdsThemeProvider.js to
src/runtime/engines/core-default/WpdsThemeProvider.js so the kernel
no longer holds a DS-specific provider in its tree. core-default
re-exports it via EngineSource.ThemeProvider; core-single-pane reuses
the same provider by importing from its sibling engine (single-pane
shares the WPDS contract).

Pure relocation — no behavior change yet. Follow-up commits drop the
kernel-side import + add the neutral-wrapper fallback.
